move run_uvsim tests to new file
