+++ pyuvsim/tests/test_mpi_uvsim.py
@@ -5,7 +5,6 @@
 import os
 
 import numpy as np
-import yaml
 import resource
 import time
 import pytest
@@ -18,14 +17,12 @@ from mpi4py import MPI
 from astropy import units
 from astropy.coordinates import Latitude, Longitude
 
-from pyuvdata import UVData
 from pyuvdata.data import DATA_PATH
 
 from pyuvsim.data import DATA_PATH as SIM_DATA_PATH
 import pyuvsim
 from pyuvsim import mpi
 from pyuvsim.astropy_interface import Time
-import pyuvsim.tests as simtest
 import pyradiosky
 
 
@@ -62,94 +59,6 @@ def fake_tasks():
 
 def test_mpi_version():
     assert MPI.VERSION == 3
-
-
-def test_run_uvsim(cst_beam):
-    hera_uv = UVData()
-    hera_uv.read_uvfits(EW_uvfits_file)
-    beam = cst_beam
-    beamfile = os.path.join(simtest.TESTDATA_PATH, "temp.uvbeam")
-    beam.write_beamfits(beamfile)
-    beam_list = pyuvsim.BeamList([beamfile])
-    mock_keywords = {"Nsrcs": 3}
-
-    with pytest.raises(TypeError, match='input_uv must be UVData object'):
-        pyuvsim.run_uvdata_uvsim('not_uvdata', beam_list, quiet=True)
-
-    mpi.start_mpi()
-    catalog, mock_kwds = pyuvsim.simsetup.create_mock_catalog(
-        hera_uv.time_array[0], return_data=True, **mock_keywords
-    )
-    uv_out = pyuvsim.run_uvdata_uvsim(hera_uv, beam_list, catalog=catalog)
-    rank = mpi.get_rank()
-    if rank == 0:
-        assert np.allclose(uv_out.data_array, hera_uv.data_array, atol=5e-3)
-    os.remove(beamfile)
-
-
-@pytest.mark.filterwarnings("ignore:The frequency field is included in the recarray")
-def test_run_paramfile_uvsim():
-    # Test vot and txt catalogs for parameter simulation
-
-    uv_ref = UVData()
-    uv_ref.read_uvfits(os.path.join(SIM_DATA_PATH, 'testfile_singlesource.uvfits'))
-    uv_ref.unphase_to_drift(use_ant_pos=True)
-
-    param_filename = os.path.join(SIM_DATA_PATH, 'test_config', 'param_1time_1src_testcat.yaml')
-    with open(param_filename) as pfile:
-        params_dict = yaml.safe_load(pfile)
-    tempfilename = params_dict['filing']['outfile_name']
-
-    # This test obsparam file has "single_source.txt" as its catalog.
-    pyuvsim.uvsim.run_uvsim(param_filename)
-
-    uv_new_txt = UVData()
-    with pytest.warns(UserWarning) as antdiam:
-        uv_new_txt.read_uvfits(tempfilename)
-    assert str(antdiam.pop().message).startswith('antenna_diameters is not set')
-    uv_new_txt.unphase_to_drift(use_ant_pos=True)
-    os.remove(tempfilename)
-
-    param_filename = os.path.join(SIM_DATA_PATH, 'test_config', 'param_1time_1src_testvot.yaml')
-    pyuvsim.uvsim.run_uvsim(param_filename)
-
-    uv_new_vot = UVData()
-    with pytest.warns(UserWarning) as antdiam:
-        uv_new_vot.read_uvfits(tempfilename)
-    assert str(antdiam.pop().message).startswith('antenna_diameters is not set')
-    uv_new_vot.unphase_to_drift(use_ant_pos=True)
-    os.remove(tempfilename)
-    uv_new_txt.history = uv_ref.history  # History includes irrelevant info for comparison
-    uv_new_vot.history = uv_ref.history
-    uv_new_txt.object_name = uv_ref.object_name
-    uv_new_vot.object_name = uv_ref.object_name
-    assert uv_new_txt == uv_ref
-    assert uv_new_vot == uv_ref
-
-
-@pytest.mark.filterwarnings("ignore:The frequency field is included in the recarray")
-def test_run_paramdict_uvsim():
-    # Running a simulation from parameter dictionary.
-
-    params = pyuvsim.simsetup._config_str_to_dict(
-        os.path.join(SIM_DATA_PATH, 'test_config', 'param_1time_1src_testcat.yaml')
-    )
-
-    pyuvsim.run_uvsim(params, return_uv=True)
-
-
-@pytest.mark.parametrize(
-    "spectral_type",
-    ["flat", "subband", "spectral_index"])
-def test_run_gleam_uvsim(spectral_type):
-    params = pyuvsim.simsetup._config_str_to_dict(
-        os.path.join(SIM_DATA_PATH, 'test_config', 'param_1time_1src_testgleam.yaml')
-    )
-    params["sources"]["spectral_type"] = spectral_type
-    params["sources"].pop("min_flux")
-    params["sources"].pop("max_flux")
-
-    pyuvsim.run_uvsim(params, return_uv=True)
 
 
 def test_mpi_funcs():
@@ -271,29 +180,6 @@ def test_big_bcast_gather_loop(fake_tasks):
     gathered = mpi.big_gather(mpi.world_comm, broadcast, root=0, MAX_BYTES=27)
 
     assert broadcast == gathered[0]
-
-
-@pytest.mark.skipif('not pyuvsim.astropy_interface.hasmoon')
-def test_sim_on_moon():
-    from pyuvsim.astropy_interface import MoonLocation
-    param_filename = os.path.join(SIM_DATA_PATH, 'test_config', 'obsparam_tranquility_hex.yaml')
-    param_dict = pyuvsim.simsetup._config_str_to_dict(param_filename)
-    param_dict['select'] = {'redundant_threshold': 0.1}
-    uv_obj, beam_list, beam_dict = pyuvsim.initialize_uvdata_from_params(param_dict)
-    uv_obj.select(times=uv_obj.time_array[0])
-    tranquility_base = MoonLocation.from_selenocentric(*uv_obj.telescope_location, 'meter')
-
-    time = Time(uv_obj.time_array[0], format='jd', scale='utc')
-    sources, kwds = pyuvsim.create_mock_catalog(
-        time, array_location=tranquility_base, arrangement='zenith', Nsrcs=30, return_data=True
-    )
-    print(uv_obj.extra_keywords['world'])
-    # Run simulation.
-    uv_out = pyuvsim.uvsim.run_uvdata_uvsim(
-        uv_obj, beam_list, beam_dict, catalog=sources, quiet=True
-    )
-    assert np.allclose(uv_out.data_array[:, 0, :, 0], 0.5)
-    assert uv_out.extra_keywords['world'] == 'moon'
 
 
 def test_sharedmem_bcast_with_quantities():
